+++ CHANGELOG.md
@@ -38,6 +38,21 @@ dependencies to your build:
 
 * Make it easier to copy the stack trace from an error message [#380](https://github.com/TNG/JGiven/issues/380)
 
+# v0.18.1
+
+## Exception handling changed for TestNG
+
+When using TestNG, exceptions are not longer caught and suppressed until the end of a scenario.
+This means that steps following a failed step are no longer shown in the scenario report.
+This change was needed to ensure that TestNG reports the correct test status in case of an exception.
+
+See [PR #422](https://github.com/TNG/JGiven/pull/422) and [Issue #312](https://github.com/TNG/JGiven/issues/312) for details.
+
+## Fixed Issues
+
+* Fix issue with @Pending(executeSteps = true) that marked a scenario as failed instead of pending [#402](https://github.com/TNG/JGiven/issues/402)
+* Make TestNG ScenarioTestListener work with parallel="tests" option [#409](https://github.com/TNG/JGiven/issues/409)
+
 # v0.18.0
 
 ## New Features